Add Backbone models for the Locations in the Trip.
When adding a Location to the Trip, the server now replies with a JSON array of
the Trip Locations. Those are then put into new Backbone models and rendered to
the Trip list with a eco view template.

The template is very bare-bones right now and needs a lot of work to make it
appear like the server-generated one, but it's already way faster than the old
way (which was to update the Trip list by reloading it with a second AJAX call
that returned the HTML).