Refactor Backbone view to use a Trip view instead of Locations.
That's what we're really displaying, anyway, a Trip, not just a collection of
Locations.

To complete it, this will require the model on the server to store distance in
the DB so it can be easily serialized into JSON and passed to the Backbone
model. It's a good idea anyway.

The old Location model/collection/view is still there, but should be removed at
some point unless it's used for something else.