Add and remove events on triplocations no longer trigger rendering during load or clear.
I used two different techniques. As all model adds are handled by add_location,
I simply moved the call to render within there. It's not as simple with remove
because the removes in which we want to render happen from the triplocation
view. Therefore, since clearing is the culprit, I added a flag to disable render
during a clear.